// src/redux/user/user.types.ts
export enum UserActionTypes {
  USER_LOGIN_START = "USER_LOGIN_START",
  USER_LOGIN_SUCESS = "USER_LOGIN_SUCESS",
  USER_LOGIN_FAILURE = "USER_LOGIN_FAILURE",
  USER_LOGOUT_START = "USER_LOGOUT_START",
  USER_LOGOUT_SUCCESS = " USER_LOGOUT_SUCCESS",
  USER_LOGOUT_FAILURE = "  USER_LOGOUT_FAILURE",
  USER_UPDATE_REQUEST_START = "USER_UPDATE_REQUEST_START",
  USER_UPDATE_SUCCESS = "USER_UPDATE_SUCCESS",
  USER_UPDATE_FAILURE = "USER_UPDATE_FAILURE",
  USER_UPDATE_PROFILE_RESET = "USER_UPDATE_PROFILE_RESET",
  HYDRATE = "__NEXT_REDUX_WRAPPER_HYDRATE__",
}